Now time is also going in nanosecond as double.

+++ inc/DataBasic.h
@@ -14,15 +14,15 @@ class DataBasic
 {
     TTree *fTree;
     Point3D fPoint;
-    unsigned long int fTime;
+    double fTime;
 
 public:
     DataBasic();
-    DataBasic(Point3D *pt, unsigned long int tm);
-    DataBasic(double x, double y, double z, unsigned long int tm);
+    DataBasic(Point3D *pt, double tm);
+    DataBasic(double x, double y, double z, double tm);
     void Fill();
-    void Fill(Point3D *pt, unsigned long int tm);
-    void Fill(double x, double y, double z, unsigned long int tm);
+    void Fill(Point3D *pt, double tm);
+    void Fill(double x, double y, double z, double tm);
     void Write();
 };
